First steps in implementing the kaiser-bessel filters

+++ hardt/libhardt/include/hbandpasskaiserbessel.h
@@ -11,7 +11,7 @@ class HBandpassKaiserBessel : public HKaiserBessel<T>
 
         /** Construct a new Kaiser-Bessel filter */
         HBandpassKaiserBessel(int fstart, int fstop, H_SAMPLE_RATE rate, int points, int att):
-            HKaiserBessel<T>(fstart, fstop, (float) rate, points, att) {}
+            HKaiserBessel<T>(fstart, fstop, rate, points, att) {}
 };
 
 #endif
