+++ hardt/libhardt/include/hhighpasskaiserbessel.h
@@ -11,7 +11,7 @@ class HHighpassKaiserBessel : public HKaiserBessel<T>
 
         /** Construct a new Kaiser-Bessel filter */
         HHighpassKaiserBessel(int fstart, H_SAMPLE_RATE rate, int points, int att):
-            HKaiserBessel<T>(fstart, rate / 2, (float) rate, points, att) {}
+            HKaiserBessel<T>(fstart, rate / 2, rate, points, att) {}
 };
 
 #endif
